scaffolding for argument parsing
- embed git and cargo version information in executable
- import clap to handle argument parsing

outline the tree of subcommands
===============================

`analyze` is intended to parse the puzzle hunt graph and display
helpful information on how bindery will interpret the input file.

`render` will take the same input file as analyze but will produce the
c or js versions for the specified target platform instead of
presenting introspective information.

`build` will manage the build process, taking the output from `render`
and sending it through the build pipeline for the target architecture.